requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg2==2.8.6
drf-spectacular==0.27.0
requests==2.31.0